+++ test/twg2/dependency/test/ClasspathReplaceTest.java
@@ -7,8 +7,6 @@ import java.io.UnsupportedEncodingException;
 
 import javax.xml.transform.TransformerException;
 
-import lombok.val;
-
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -18,16 +16,17 @@ public class ClasspathReplaceTest {
 
 	@Test
 	public void addDependencyTest() throws UnsupportedEncodingException, TransformerException {
-		val jcollInterfacesCp = ClasspathExampleFiles.getJCollectionUtil("no", ClasspathExampleFiles.getJCollectionInterfacesEntry());
+		String jcollInterfacesCp = ClasspathExampleFiles.getJCollectionUtil("no", ClasspathExampleFiles.nwln, "    ",
+				ClasspathExampleFiles.getJCollectionInterfacesEntry());
 
-		val jcollInterfacesCpPlain = ClasspathExampleFiles.getJCollectionUtil("no");
-		val doc = EclipseClasspathDoc.fromXml(new File(""), new ByteArrayInputStream(jcollInterfacesCpPlain.getBytes("utf-8")));
+		String jcollInterfacesCpPlain = ClasspathExampleFiles.getJCollectionUtil("no", "", "");
+		EclipseClasspathDoc doc = EclipseClasspathDoc.fromXml(new File(""), new ByteArrayInputStream(jcollInterfacesCpPlain.getBytes("utf-8")));
 
 		doc.addClassPathEntry(ClasspathExampleFiles.getJCollectionInterfacesEntry());
 
-		val output = new ByteArrayOutputStream();
+		ByteArrayOutputStream output = new ByteArrayOutputStream();
 		doc.toXml(output);
-		val resStr = new String(output.toByteArray(), "utf-8");
+		String resStr = new String(output.toByteArray(), "utf-8");
 
 		Assert.assertEquals(jcollInterfacesCp.trim(), resStr.trim());
 	}